backend/requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
fastapi
pydantic
python-dotenv